refactor: 헥사고날 아키텍처 적용 (#153)
* refactor: code convention 적용

* refactor: 단위, 통합 테스트 패키지 통합
* refactor: 도메인 패키지 구조 변경
* refactor: ci build 시 통합 테스트 수행
* refactor: 매핑 정보 분리
* test: 테스트 컨테이너 설정 변경
* refactor: 서비스 mocking 테스트 제거

* refactor: 지역 인증 api 위치 수정

* refactor: area ar 정의
* refactor: infra mail 구조 개선
* refactor: infra cache 구조 개선
* refactor: security 계층 분리
* refactor: infra 파일 storage 구조 개선
* refactor: 관심사 도메인 분리
* refactor: book 도메인 구조 개선
* refactor: bookcase 도메인 분리
* refactor: member ar 개선
* refactor: category 도메인 분리
* refactor: post ar 개선
* refactor: file ar 개선
* refactor: notification 구조 개선
* refactor: trade ar 개선
* refactor: chat ar 개선

* refactor: style 적용 및 네이밍 수정

+++ src/main/java/com/bob/global/utils/web/CookieUtils.java
@@ -1,49 +1,51 @@
 package com.bob.global.utils.web;
 
-import jakarta.servlet.http.Cookie;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
 import java.util.Arrays;
 import java.util.Objects;
 import java.util.Optional;
+
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
+
 import lombok.Setter;
+
 import org.springframework.http.ResponseCookie;
 
 public class CookieUtils {
 
-  @Setter
-  private static String sameSite = "LAX";
+    @Setter
+    private static String sameSite = "LAX";
 
-  public static String getCookie(HttpServletRequest request, String name) {
-    return Optional.ofNullable(request.getCookies())
-        .flatMap(cookies -> Arrays.stream(cookies)
-            .filter(cookie -> Objects.equals(name, cookie.getName()))
-            .map(Cookie::getValue)
-            .findAny()).orElse(null);
-  }
+    public static String getCookie(HttpServletRequest request, String name) {
+        return Optional.ofNullable(request.getCookies())
+            .flatMap(cookies -> Arrays.stream(cookies)
+                .filter(cookie -> Objects.equals(name, cookie.getName()))
+                .map(Cookie::getValue)
+                .findAny()).orElse(null);
+    }
 
-  public static void addCookie(HttpServletResponse response, String name, String value, int maxAge) {
-    ResponseCookie cookie = ResponseCookie.from(name, value)
-        .path("/")
-        .sameSite(sameSite)
-        .httpOnly(true)
-        .secure(true)
-        .maxAge(maxAge)
-        .build();
+    public static void addCookie(HttpServletResponse response, String name, String value, Long maxAge) {
+        ResponseCookie cookie = ResponseCookie.from(name, value)
+            .path("/")
+            .sameSite(sameSite)
+            .httpOnly(true)
+            .secure(true)
+            .maxAge(maxAge)
+            .build();
 
-    response.addHeader("Set-Cookie", cookie.toString());
-  }
+        response.addHeader("Set-Cookie", cookie.toString());
+    }
 
-  public static void removeCookie(HttpServletResponse response, String name) {
-    ResponseCookie cookie = ResponseCookie.from(name, "")
-        .path("/")
-        .sameSite(sameSite)
-        .httpOnly(true)
-        .secure(true)
-        .maxAge(0)
-        .build();
+    public static void removeCookie(HttpServletResponse response, String name) {
+        ResponseCookie cookie = ResponseCookie.from(name, "")
+            .path("/")
+            .sameSite(sameSite)
+            .httpOnly(true)
+            .secure(true)
+            .maxAge(0)
+            .build();
 
-    response.addHeader("Set-Cookie", cookie.toString());
-  }
+        response.addHeader("Set-Cookie", cookie.toString());
+    }
 }
-
